pyproject.toml
[tool.poetry]
name = "user-profile"
version = "0.1.0"
description = ""
authors = ["specnazm <laketic.milena98@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.1.2"
Flask-SQLAlchemy = "^2.5.1"
Flask-Migrate = "^3.1.0"
psycopg2-binary = "^2.9.3"
SQLAlchemy-serializer = "^1.4.1"
kafka-python = "^2.0.2"
PyJWT = "^2.4.0"
[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"